chore: Move sniping into it's own file
